+++ src/renderer/render/media/audio.ts
@@ -1,11 +1,12 @@
 ﻿import { cache } from '../../../parser/inline/media';
 import { html } from 'typed-dom';
 
-export function audio(url: string, alt: string): HTMLAudioElement {
-  return cache.has(url)
-    ? cache.get(url)!.cloneNode(true) as HTMLAudioElement
-    : cache.set(url, html('audio',{
-        src: url,
+export function audio(url: URL, alt: string): HTMLAudioElement | undefined {
+  if (!['.oga', '.ogg'].includes(url.pathname.split(/(?=\.)/).pop()!)) return;
+  return cache.has(url.href)
+    ? cache.get(url.href)!.cloneNode(true) as HTMLAudioElement
+    : cache.set(url.href, html('audio',{
+        src: url.href,
         alt,
         controls: '',
         style: 'width: 100%;',