+++ src/renderer/render/media/gist.ts
@@ -3,13 +3,14 @@ import { cache } from '../../../parser/inline/media';
 import { sanitize } from 'dompurify';
 import DOM, { html } from 'typed-dom';
 
-export function gist(url: string): HTMLElement {
-  if (cache.has(url)) return cache.get(url)!.cloneNode(true);
+export function gist(url: URL): HTMLElement | undefined {
+  if (!['https://gist.github.com'].includes(url.origin)) return;
+  if (cache.has(url.href)) return cache.get(url.href)!.cloneNode(true);
   return DOM.div({
     style: 'position: relative;',
-  }, [DOM.em(`loading ${url}`)], () => {
+  }, [DOM.em(`loading ${url.href}`)], () => {
     const outer = html('div');
-    void $.ajax(`${url}.json`, {
+    void $.ajax(`${url.href}.json`, {
       dataType: 'jsonp',
       timeout: 10 * 1e3,
       cache: true,
@@ -20,10 +21,10 @@ export function gist(url: string): HTMLElement {
         void gist.insertBefore(
           DOM.div({ class: 'gist-description' }, [
             DOM.a({ style: 'color: #555; font-size: 14px; font-weight: 600;' }, description, () =>
-              parse(escape(url)).querySelector('a')!),
+              parse(escape(url.href)).querySelector('a')!),
           ]).element,
           gist.firstChild);
-        void cache.set(url, outer.cloneNode(true));
+        void cache.set(url.href, outer.cloneNode(true));
         if (document.head.querySelector(`link[rel="stylesheet"][href="${stylesheet}"]`)) return;
         void document.head.appendChild(html('link', {
           rel: 'stylesheet',
@@ -33,7 +34,7 @@ export function gist(url: string): HTMLElement {
       },
       error({ status, statusText }) {
         assert(Number.isSafeInteger(status));
-        outer.innerHTML = parse(`*${escape(url)}\\\n-> ${status}: ${escape(statusText)}*`).querySelector('p')!.innerHTML;
+        outer.innerHTML = parse(`*${escape(url.href)}\\\n-> ${status}: ${escape(statusText)}*`).querySelector('p')!.innerHTML;
       },
     });
     return outer;