+++ src/renderer/render/media/twitter.ts
@@ -12,17 +12,18 @@ declare global {
 let widgetScriptRequested = !!window.twttr;
 const cache = new Cache<string, HTMLElement>(100);
 
-export function twitter(url: string): HTMLElement {
-  if (cache.has(url)) {
-    const el = cache.get(url)!.cloneNode(true);
+export function twitter(url: URL): HTMLElement | undefined {
+  if (!['https://twitter.com'].includes(url.origin)) return;
+  if (cache.has(url.href)) {
+    const el = cache.get(url.href)!.cloneNode(true);
     window.twttr && void window.twttr.widgets.load(el);
     return el;
   }
   return DOM.div({
     style: 'position: relative;',
-  }, [DOM.em(`loading ${url}`)], () => {
+  }, [DOM.em(`loading ${url.href}`)], () => {
     const outer = html('div');
-    void $.ajax(`https://publish.twitter.com/oembed?url=${url.replace('?', '&')}`, {
+    void $.ajax(`https://publish.twitter.com/oembed?url=${url.href.replace('?', '&')}`, {
       dataType: 'jsonp',
       timeout: 10 * 1e3,
       cache: true,
@@ -30,7 +31,7 @@ export function twitter(url: string): HTMLElement {
         outer.innerHTML = sanitize(`<div style="margin-top: -10px; margin-bottom: -10px;">${html}</div>`, { ADD_TAGS: ['script'] });
         const script = outer.querySelector('script');
         script && void script.remove();
-        void cache.set(url, outer.cloneNode(true));
+        void cache.set(url.href, outer.cloneNode(true));
         if (window.twttr) return void window.twttr.widgets.load(outer);
         if (widgetScriptRequested || !script) return;
         widgetScriptRequested = true;
@@ -43,7 +44,7 @@ export function twitter(url: string): HTMLElement {
       },
       error({ status, statusText }) {
         assert(Number.isSafeInteger(status));
-        outer.innerHTML = parse(`*${escape(url)}\\\n-> ${status}: ${escape(statusText)}*`).querySelector('p')!.innerHTML;
+        outer.innerHTML = parse(`*${escape(url.href)}\\\n-> ${status}: ${escape(statusText)}*`).querySelector('p')!.innerHTML;
       },
     });
     return outer;