+++ src/renderer/render/media/video.ts
@@ -1,11 +1,12 @@
 ﻿import { cache } from '../../../parser/inline/media';
 import { html } from 'typed-dom';
 
-export function video(url: string, alt: string): HTMLVideoElement {
-  return cache.has(url)
-    ? cache.get(url)!.cloneNode(true) as HTMLVideoElement
-    : cache.set(url, html('video', {
-        src: url,
+export function video(url: URL, alt: string): HTMLVideoElement | undefined {
+  if (!['.webm', '.ogv'].includes(url.pathname.split(/(?=\.)/).pop()!)) return;
+  return cache.has(url.href)
+    ? cache.get(url.href)!.cloneNode(true) as HTMLVideoElement
+    : cache.set(url.href, html('video', {
+        src: url.href,
         alt,
         muted: '',
         controls: '',