+++ src/renderer/render/media/youtube.ts
@@ -1,6 +1,7 @@
 ﻿import DOM from 'typed-dom';
 
-export function youtube(url: string): HTMLElement {
+export function youtube(url: URL): HTMLElement | undefined {
+  if (!['https://www.youtube.com', 'https://youtu.be'].includes(url.origin)) return;
   return DOM.div({
     style: 'position: relative;',
   }, [
@@ -9,8 +10,8 @@ export function youtube(url: string): HTMLElement {
     }, [
       DOM.iframe({
         src: `https://www.youtube.com/embed/${
-          url.startsWith('https://www.youtube.com/') && url.replace(/.+?=/, '').replace(/&/, '?') ||
-          url.startsWith('https://youtu.be/') && url.slice(url.indexOf('/', 9) + 1)
+          url.origin === 'https://www.youtube.com' && url.href.replace(/.+?=/, '').replace(/&/, '?') ||
+          url.origin === 'https://youtu.be' && url.href.slice(url.href.indexOf('/', 9) + 1)
         }`,
         allowfullscreen: '',
         frameborder: '0',